requirements.txt
backports.functools-lru-cache==1.6.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cycler==0.10.0
Django==1.11.29
django-crontab==0.7.1
django-filter==1.1.0
djangorestframework==3.9.4
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.1.0
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.5
mechanize==0.4.5
mysqlclient==1.4.6
numpy==1.16.6
pandas==0.24.2
Pillow==6.2.2
psycopg2==2.7.3.2
pyaml==20.4.0
pygraphviz==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytidylib==0.3.2
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.2.3
simplejson==3.17.0
six==1.15.0
subprocess32==3.5.4
uritemplate==3.0.1
urllib3==1.25.9
uTidylib==0.6
webencodings==0.5.1
